+++ komapper-core/src/main/kotlin/org/komapper/core/dsl/operator/CriteriaContext.kt
@@ -1,6 +1,5 @@
 package org.komapper.core.dsl.operator
 
-import org.komapper.core.dsl.expression.EscapeExpression
 import org.komapper.core.dsl.expression.SqlBuilderScope
 
 /**
@@ -14,14 +13,4 @@ interface CriteriaContext {
      * @param build the SQL builder
      */
     fun add(build: SqlBuilderScope.() -> Unit)
-
-    /**
-     * Does not escape the given string.
-     */
-    fun <S : CharSequence> text(value: S): EscapeExpression
-
-    /**
-     * Escapes the given string.
-     */
-    fun <S : CharSequence> escape(value: S): EscapeExpression
 }
